+++ tests/test_pipeline.py
@@ -1555,8 +1555,12 @@ class TestPipeAnalysis:
         new_image_row = dummy_PipeAnalysis.images.iloc[0]
         new_image_row.name = 10
 
-        dummy_PipeAnalysis.images = dummy_PipeAnalysis.images.append(
-            new_image_row
+        dummy_PipeAnalysis.images = pd.concat(
+            [
+                dummy_PipeAnalysis.images,
+                # need to transpose the series to a dataframe to concat
+                new_image_row.to_frame().T.set_index('name')
+            ]
         )
 
         dummy_PipeAnalysis.images.loc[10, 'skyreg_id'] = 4